Repairing some things that weren't working
Alt-D sort of works to take the focus to the addressbar. It looks like it works only when the focus is not in the web browser but rather on the form in general, though MS documentation says it must be picked from anywhere.

Close works as another minimize now - which is very useful if you're used to Alt-Space-C like I am, but don't want this app to exit when you do that, like I do. I guess you could see a pattern about feature choices there (obviously).
Also, Escape now works as minimize, so you could copy a word, Ctrl-Alt-D to see its definition, then Esc to minimize the definition. Easy, ain't it? :)
So what do you do to stop a page from loading, if Esc won't do that? Well, supposedly, getting your focus to the addressbar will stop loading the current page (assuming you want to change the URL and so don't need the current page), but this doesn't seem to work yet.

Also including Release version of binaries with this commit, just in case someone needs them (haha).